fix(ngAria): bind to `keydown` instead of `keypress` in `ngClick`
The ngAria keyboard support for html elements used as button now binds to the keydown
event instead of keypress to better emulate behavior of native buttons.

BREAKING CHANGE: The ngAria configuration flag that controls binding to keyboard input
has been renamed from `bindKeypress` to `bindKeydown` so that it accurately dscribes the
implemented behavior.

Closes #14063